2022_12_super_bomber_elf/a0.py
from websocket import create_connection

ws = None
try:
    ws = create_connection("ws://172.16.173.86:8080/")
    while True:
        print(ws.recv())

        command = input("enter command: ")
        ws.send(command)
finally:
    if ws:
        ws.close()
