requirements.txt
async-timeout==4.0.2
black==22.12.0
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
Deprecated==1.2.13
idna==3.3
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.10.3
platformdirs==2.6.0
pyparsing==3.0.9
readchar==4.0.3
redis==4.3.4
requests==2.28.1
six==1.16.0
spotipy==2.20.0
tomli==2.0.1
urllib3==1.26.12
websocket-client==1.4.2
windows-curses==2.3.1
wrapt==1.14.1
